+++ src/ska_tmc_common/test_helpers/helper_dish_ln_device.py
@@ -55,7 +55,7 @@ class HelperDishLNDevice(HelperBaseDevice):
         self._offset: dict = {"off_xel": 0.0, "off_el": 0.0}
         self._actual_pointing: list = []
         self._kvalue: int = 0
-        self._isSubsystemAvailable = False
+        self._isSubsystemAvailable = True
         self._dish_kvalue_validation_result = str(int(ResultCode.STARTED))
         self._dish_mode = DishMode.STANDBY_LP
         self._pointing_state = PointingState.NONE
@@ -341,13 +341,9 @@ class HelperDishLNDevice(HelperBaseDevice):
         else:
             self._dish_kvalue_validation_result = str(int(ResultCode.UNKNOWN))
 
-        self.logger.info(
-            "Push Dish kvalue Validation Result as %s",
-            self._dish_kvalue_validation_result,
-        )
         self.push_change_event(
             "kValueValidationResult",
-            str(int(self._dish_kvalue_validation_result)),
+            self._dish_kvalue_validation_result,
         )
 
     def push_dish_kvalue_val_result_after_initialization(self):
@@ -514,7 +510,6 @@ class HelperDishLNDevice(HelperBaseDevice):
             return self.induce_fault("SetStandbyFPMode")
         if self.dev_state() != DevState.STANDBY:
             self.set_state(DevState.STANDBY)
-            time.sleep(0.1)
             self.push_change_event("State", self.dev_state())
         # Set the Dish Mode
         self.set_dish_mode(DishMode.STANDBY_FP)
@@ -562,7 +557,6 @@ class HelperDishLNDevice(HelperBaseDevice):
         # Set the device state
         if self.dev_state() != DevState.STANDBY:
             self.set_state(DevState.STANDBY)
-            time.sleep(0.1)
             self.push_change_event("State", self.dev_state())
 
         # Set the Dish ModeLP
@@ -611,7 +605,6 @@ class HelperDishLNDevice(HelperBaseDevice):
         # Set the device state
         if self.dev_state() != DevState.ON:
             self.set_state(DevState.ON)
-            time.sleep(0.1)
             self.push_change_event("State", self.dev_state())
         # Set the pointing state
         if self._pointing_state != PointingState.READY:
@@ -664,7 +657,6 @@ class HelperDishLNDevice(HelperBaseDevice):
         # Set device state
         if self.dev_state() != DevState.DISABLE:
             self.set_state(DevState.DISABLE)
-            time.sleep(0.1)
             self.push_change_event("State", self.dev_state())
 
         # Set Dish Mode
@@ -1088,7 +1080,7 @@ class HelperDishLNDevice(HelperBaseDevice):
     )
     def EndScan(self) -> Tuple[List[ResultCode], List[str]]:
         """
-        This method updates the scanID attribute of Dish Master
+        This method invokes EndScan command on Dish Master
         :return: ResultCode and message
         :rtype: Tuple[List[ResultCode], List[str]]
         """
@@ -1096,8 +1088,6 @@ class HelperDishLNDevice(HelperBaseDevice):
         self.update_command_info(END_SCAN)
         if self.defective_params["enabled"]:
             return self.induce_fault("EndScan")
-            # On Real Dish Leaf Node the scanID attribute of Dish Master
-            # is getting updated
             # TBD: Add your dish mode change logic here if required
         return ([ResultCode.OK], [""])