+++ src/ska_tmc_common/test_helpers/helper_mccs_controller_device.py
@@ -121,7 +121,7 @@ class HelperMCCSController(HelperBaseDevice):
             thread = threading.Timer(
                 self._delay,
                 function=self.push_command_result,
-                args=[command_id, result, fault_message],
+                args=[result, command_id, fault_message],
             )
             thread.start()
             return [ResultCode.QUEUED], [command_id]
@@ -157,8 +157,6 @@ class HelperMCCSController(HelperBaseDevice):
     def wait_and_update_exception(self, command_id):
         """Waits for 5 secs before pushing a longRunningCommandResult event."""
         with EnsureOmniThread():
-            time.sleep(5)
-
             command_result = (
                 command_id,
                 json.dumps(
@@ -184,7 +182,7 @@ class HelperMCCSController(HelperBaseDevice):
                 "Sleep %s for command %s ", delay_value, command_name
             )
 
-        self.push_command_result(command_id, ResultCode.OK)
+        self.push_command_result(ResultCode.OK, command_id)
 
     def is_Allocate_allowed(self) -> bool:
         """
@@ -227,8 +225,10 @@ class HelperMCCSController(HelperBaseDevice):
             return self.induce_fault("Allocate", command_id)
         if self._raise_exception:
             self.logger.info("exception thread")
-            thread = threading.Thread(
-                target=self.wait_and_update_exception, args=[command_id]
+            thread = threading.Timer(
+                interval=5,
+                function=self.wait_and_update_exception,
+                args=[command_id],
             )
             thread.start()
             return [ResultCode.QUEUED], [command_id]
@@ -289,8 +289,10 @@ class HelperMCCSController(HelperBaseDevice):
 
         if self._raise_exception:
             self.logger.info("exception thread")
-            thread = threading.Thread(
-                target=self.wait_and_update_exception, args=[command_id]
+            thread = threading.Timer(
+                interval=5,
+                function=self.wait_and_update_exception,
+                args=[command_id],
             )
             thread.start()
             return [ResultCode.QUEUED], [command_id]
@@ -349,8 +351,10 @@ class HelperMCCSController(HelperBaseDevice):
 
         if self._raise_exception:
             self.logger.info("exception thread")
-            thread = threading.Thread(
-                target=self.wait_and_update_exception, args=[command_id]
+            thread = threading.Timer(
+                interval=5,
+                function=self.wait_and_update_exception,
+                args=[command_id],
             )
             thread.start()
             return [ResultCode.QUEUED], [command_id]